// src/types/types.ts
export interface SizeTableData {
  id: string;
  title: string;
  sizes: string[];
  length: number[];
  chest: number[];
}

export interface PriceTableData {
  tshirt: number[];
  hoodie: number[];
  titles: string[];
}

export interface ContactInfo {
  name: string;
  phone: string;
  instagram: string;
  facebook: string;
}